Added Player

+++ src/com/zzstudios/nfc/entity/Entity.java
@@ -5,7 +5,7 @@ import com.zzstudios.nfc.input.InputHandler;
 import com.zzstudios.nfc.level.Level;
 
 public class Entity {
-	public int x, y, rotation;
+	public double x, y, rotation;
 	
 	public int xSize, ySize;
